Apis labenusers atualizado

+++ modulo2/apis-labenusers/src/App.js
@@ -1,43 +1,22 @@
-import axios from "axios"
 import React from "react"
+import axios from "axios"
+import Cadastro from "./telas/Cadastro"
+import ListaDeUsuarios from "./telas/ListaDeUsuarios"
+import DetalhesDoUsuario from "./telas/DetalhesDoUsuario"
 
 class App extends React.Component {
   state = {
-    inputNome: "",
-    inputEmail: "",
+    tela: "cadastro",
+    usuario: {},
     listaDeUsuarios: [],
-    telaHome: true
   }
 
-  componentDidMount() {
-    this.getAllUsers()
+  irParaTelaCadastro = () => {
+    this.setState({ tela: "cadastro" })
   }
 
-  onChangeNome = (event) => {
-    this.setState({ inputNome: event.target.value })
-  }
-
-  onChangeEmail = (event) => {
-    this.setState({ inputEmail: event.target.value })
-  }
-
-  createUser = () => {
-    const url = "https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users"
-    const body = {
-      name: this.state.inputNome,
-      email: this.state.inputEmail
-    }
-    const axiosConfig = { headers: { Authorization: "sofia-sesso-vaughan" } }
-
-    axios.post(url, body, axiosConfig)
-      .then((response) => {
-        alert("Usuário cadastrado!")
-        this.setState({ inputNome: "", inputEmail: "" })
-        this.getAllUsers()
-      })
-      .catch((error) => {
-        alert(error.response.data.message)
-      })
+  irParaTelaLista = () => {
+    this.setState({ tela: "lista" })
   }
 
   getAllUsers = () => {
@@ -53,50 +32,61 @@ class App extends React.Component {
       })
   }
 
-  deleteUser = (idDoUsuarioClicado) => {
-    const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${idDoUsuarioClicado}`
+  getUserById = (id) => {
+    const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`
     const axiosConfig = { headers: { Authorization: "sofia-sesso-vaughan" } }
 
-    axios.delete(url, axiosConfig)
+    axios.get(url, axiosConfig)
       .then((response) => {
-        alert("Usuário deletado!")
-        this.getAllUsers()
+        this.setState({ usuario: response.data, tela: "detalhes" })
       })
       .catch((error) => {
-        console.log(error.response.data)
         alert(error.response.data)
       })
   }
 
-  irParaTelaLista = () => {
-    this.setState({ telaHome: false })
-  }
+  deleteUser = (id) => {
+    const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`
+    const axiosConfig = { headers: { Authorization: "sofia-sesso-vaughan" } }
 
-  voltarParatelaHome = () => {
-    this.setState({ telaHome: true })
+    if (window.confirm("Tem certeza de que deseja deletar?")) {
+      axios.delete(url, axiosConfig)
+        .then(() => {
+          alert("Usuário deletado!")
+          this.getAllUsers()
+          this.setState({tela: "lista"})
+        })
+        .catch((error) => {
+          alert(error.response.data)
+        })
+    }
   }
 
   render() {
-    const listaDeNomes = this.state.listaDeUsuarios.map((usuario) => {
-      return (
-        <ul>
-          <li>{usuario.name} <button onClick={() => this.deleteUser(usuario.id)}>Deletar Usuário</button></li>
-        </ul>)
-    })
 
     const renderizaTela = () => {
-      if (this.state.telaHome === true) {
-        return <div>
-          <button onClick={this.irParaTelaLista}>Ver Lista de Nomes</button>
-          <input value={this.state.inputNome} onChange={this.onChangeNome} placeholder="Nome" />
-          <input value={this.state.inputEmail} onChange={this.onChangeEmail} placeholder="E-mail" />
-          <button onClick={this.createUser}>Criar Usuário</button>
-        </div>
-      } else {
-        return <div>
-          <button onClick={this.voltarParatelaHome}>Voltar</button>
-          {listaDeNomes}
-        </div>
+      switch (this.state.tela) {
+        case "cadastro":
+          return <Cadastro
+            irParaTelaLista={this.irParaTelaLista}
+          />
+        case "lista":
+          return <ListaDeUsuarios
+            getUserById={this.getUserById}
+            irParaTelaCadastro={this.irParaTelaCadastro}
+            deleteUser={this.deleteUser}
+            getAllUsers={this.getAllUsers}
+            listaDeUsuarios={this.state.listaDeUsuarios}
+          />
+        case "detalhes":
+          return <DetalhesDoUsuario
+            usuario={this.state.usuario}
+            irParaTelaCadastro={this.irParaTelaCadastro}
+            irParaTelaLista={this.irParaTelaLista}
+            deleteUser={this.deleteUser}
+            getAllUsers={this.getAllUsers}
+            getUserById={this.getUserById}
+          />
       }
     }
